Continued refactoring to match modern idioms; shifted Pandoc from a tuple struct to a c-type one
