+++ Cargo.toml
@@ -12,6 +12,5 @@ include = ["Cargo.toml", "src/**/*.rs", "tests/**/*.rs", "examples/**/*.rs", "RE
 edition = "2018"
 
 [dependencies]
-serde = "1.0"
-serde_derive = "1.0"
+serde = {version = "1.0", features=["derive"]}
 serde_json = "1.0"
