+++ examples/definition.rs
@@ -1,27 +1,27 @@
-extern crate serde_json;
-
-extern crate pandoc_types;
-
-use pandoc_types::definition::*;
+use pandoc_types::*;
+use std::collections::HashMap;
 
 fn main() {
-    let mut meta = Meta::null();
-    meta.0.insert(
+    let mut meta = HashMap::new();
+    meta.insert(
         "title".to_owned(),
         MetaValue::MetaInlines(vec![Inline::Str("a".to_owned())]),
     );
 
-    let doc = Pandoc(
+    let blocks = vec![
+        Block::Header(
+            1,
+            Attr("a".to_owned(), vec![], vec![]),
+            vec![Inline::Str("a".to_owned())],
+        ),
+        Block::Para(vec![Inline::Str("b".to_owned())]),
+    ];
+
+    let doc = Pandoc {
         meta,
-        vec![
-            Block::Header(
-                1,
-                Attr("a".to_owned(), vec![], vec![]),
-                vec![Inline::Str("a".to_owned())],
-            ),
-            Block::Para(vec![Inline::Str("b".to_owned())]),
-        ],
-    );
+        blocks,
+        api_version: None
+    };
 
     let s = serde_json::to_string(&doc).unwrap();
     println!("serialized = {}", s);
